Выполнено Финальное задание 22

+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/order/OrderDto.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.experimental.FieldDefaults;
 
+import java.math.BigDecimal;
 import java.util.Map;
 import java.util.UUID;
 
@@ -21,7 +22,7 @@ public class OrderDto {
     Double deliveryWeight;
     Double deliveryVolume;
     boolean fragile;
-    Double totalPrice;
-    Double deliveryPrice;
-    Double productPrice;
+    BigDecimal totalPrice;
+    BigDecimal deliveryPrice;
+    BigDecimal productPrice;
 }
